Title: 사각형 채우기 3, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241009/사각형 채우기 3/rectangle-fill-3.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-int dp[MAX_SIZE];
+long long dp[MAX_SIZE];
 int N;
 
 int count(int col) {
